requirements.txt
django
django_taggit
django_formtools
fuzzywuzzy
Pillow
python-Levenshtein
pytz
django-redis